Correções na API
Subindo correções, dentre elas a adição do login JWT com token bearer, e correções para inserção de dados, e substituição do banco de dados usado.

+++ TG_AGM-Park/app/Config/Routes.php
@@ -7,8 +7,17 @@ use CodeIgniter\Router\RouteCollection;
  */
 $routes->get('/', 'Home::index');
 
-$routes->group('usuarios', function ($routes) {
-    $routes->post('inserir', 'Usuarios::inserir');
+// Autenticação
+$routes->post('auth/login', 'Auth::login');
+$routes->get('auth/checkauth', 'Auth::checkAuth', ['filter' => 'jwt']);
+
+// Cadastro publico para permitir login apos criar conta.
+$routes->post('usuarios/inserir', 'Usuarios::inserir');
+$routes->get('usuarios/me', 'Usuarios::me', ['filter' => 'jwt']);
+$routes->put('usuarios/me', 'Usuarios::atualizarMe', ['filter' => 'jwt']);
+
+// Apenas PROPRIETARIO pode manipular usuários
+$routes->group('usuarios', ['filter' => ['jwt', 'role:PROPRIETARIO']], function ($routes) {
     $routes->get('listar', 'Usuarios::listar');
     $routes->get('(:num)', 'Usuarios::buscar/$1');
     $routes->put('atualizar/(:num)', 'Usuarios::atualizar/$1');
@@ -23,28 +32,31 @@ $routes->group('estacionamentos', function ($routes) {
 });
 
 $routes->group('veiculos', function ($routes) {
-    $routes->post('inserir', 'Veiculos::inserir');
-    $routes->get('listar', 'Veiculos::listar');
-    $routes->put('atualizar/(:num)', 'Veiculos::atualizar/$1');
-    $routes->delete('deletar/(:num)', 'Veiculos::deletar/$1');
+    // Funcionário tem acesso a todos endpoints de veículos
+    $routes->post('inserir', 'Veiculos::inserir', ['filter' => ['jwt', 'role:PROPRIETARIO,FUNCIONARIO']]);
+    $routes->get('listar', 'Veiculos::listar', ['filter' => ['jwt', 'role:PROPRIETARIO,FUNCIONARIO']]);
+    $routes->put('atualizar/(:num)', 'Veiculos::atualizar/$1', ['filter' => ['jwt', 'role:PROPRIETARIO,FUNCIONARIO']]);
+    $routes->delete('deletar/(:num)', 'Veiculos::deletar/$1', ['filter' => ['jwt', 'role:PROPRIETARIO,FUNCIONARIO']]);
 });
 
+// Vagas: FUNCIONARIO pode ler e atualizar; inserir e deletar apenas ADMIN
 $routes->group('vagas', function ($routes) {
-    $routes->post('inserir', 'Vagas::inserir');
-    $routes->get('listar', 'Vagas::listar');
-    $routes->put('atualizar/(:num)', 'Vagas::atualizar/$1');
+    $routes->post('inserir', 'Vagas::inserir', ['filter' => ['jwt', 'role:PROPRIETARIO']]);
+    $routes->get('listar', 'Vagas::listar', ['filter' => ['jwt', 'role:PROPRIETARIO,FUNCIONARIO']]);
+    $routes->put('disponibilidade/(:num)', 'Vagas::alterarDisponibilidade/$1', ['filter' => ['jwt', 'role:PROPRIETARIO,FUNCIONARIO']]);
+    $routes->put('atualizar/(:num)', 'Vagas::atualizar/$1', ['filter' => ['jwt', 'role:PROPRIETARIO,FUNCIONARIO']]);
     $routes->put('atualizar-status-fisico/(:num)', 'Vagas::atualizarStatusFisico/$1');
-    $routes->delete('deletar/(:num)', 'Vagas::deletar/$1');
+    $routes->delete('deletar/(:num)', 'Vagas::deletar/$1', ['filter' => ['jwt', 'role:PROPRIETARIO']]);
 });
 
-$routes->group('reservas', function ($routes) {
+$routes->group('reservas', ['filter' => ['jwt', 'role:PROPRIETARIO,FUNCIONARIO']], function ($routes) {
     $routes->post('inserir', 'Reservas::inserir');
     $routes->get('listar', 'Reservas::listar');
     $routes->put('atualizar/(:num)', 'Reservas::atualizar/$1');
     $routes->delete('deletar/(:num)', 'Reservas::deletar/$1');
 });
 
-$routes->group('estadias', function ($routes) {
+$routes->group('estadias', ['filter' => ['jwt', 'role:PROPRIETARIO,FUNCIONARIO']], function ($routes) {
     $routes->post('inserir', 'Estadias::inserir');
     $routes->get('listar', 'Estadias::listar');
     $routes->put('atualizar/(:num)', 'Estadias::atualizar/$1');
@@ -59,10 +71,10 @@ $routes->group('pagamentos', function ($routes) {
 });
 
 $routes->group('FuncionarioEstacionamento', function ($routes) {
-    $routes->post('inserir', 'FuncionarioEstacionamento::inserir');
-    $routes->get('listar', 'FuncionarioEstacionamento::listar');
-    $routes->put('atualizar/(:num)', 'FuncionarioEstacionamento::atualizar/$1');
-    $routes->delete('deletar/(:num)', 'FuncionarioEstacionamento::deletar/$1');
+    $routes->post('inserir', 'FuncionarioEstacionamento::inserir', ['filter' => ['jwt', 'role:PROPRIETARIO']]);
+    $routes->get('listar', 'FuncionarioEstacionamento::listar', ['filter' => ['jwt', 'role:PROPRIETARIO,FUNCIONARIO']]);
+    $routes->put('atualizar/(:num)', 'FuncionarioEstacionamento::atualizar/$1', ['filter' => ['jwt', 'role:PROPRIETARIO']]);
+    $routes->delete('deletar/(:num)', 'FuncionarioEstacionamento::deletar/$1', ['filter' => ['jwt', 'role:PROPRIETARIO']]);
 });
 
 $routes->group('HorariosFuncionamento', function ($routes) {
@@ -72,9 +84,10 @@ $routes->group('HorariosFuncionamento', function ($routes) {
     $routes->delete('deletar/(:num)', 'HorariosFuncionamento::deletar/$1');
 });
 
+// Tarifas: FUNCIONARIO apenas leitura; alterações somente ADMIN
 $routes->group('tarifas', function ($routes) {
-    $routes->post('inserir', 'Tarifas::inserir');
-    $routes->get('listar', 'Tarifas::listar');
-    $routes->put('atualizar/(:num)', 'Tarifas::atualizar/$1');
-    $routes->delete('deletar/(:num)', 'Tarifas::deletar/$1');
+    $routes->post('inserir', 'Tarifas::inserir', ['filter' => ['jwt', 'role:PROPRIETARIO']]);
+    $routes->get('listar', 'Tarifas::listar', ['filter' => ['jwt', 'role:PROPRIETARIO,FUNCIONARIO']]);
+    $routes->put('atualizar/(:num)', 'Tarifas::atualizar/$1', ['filter' => ['jwt', 'role:PROPRIETARIO']]);
+    $routes->delete('deletar/(:num)', 'Tarifas::deletar/$1', ['filter' => ['jwt', 'role:PROPRIETARIO']]);
 });
